Split out chat boxes into separate React components
