+++ src/react-components/in-world-chat-box.js
@@ -1,0 +1,133 @@
+import React, { Component } from "react";
+import PropTypes from "prop-types";
+import classNames from "classnames";
+import styles from "../assets/stylesheets/ui-root.scss";
+import ChatCommandHelp from "./chat-command-help";
+import { faPaperPlane } from "@fortawesome/free-solid-svg-icons/faPaperPlane";
+import { faCamera } from "@fortawesome/free-solid-svg-icons/faCamera";
+import { faPlus } from "@fortawesome/free-solid-svg-icons/faPlus";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { handleTextFieldFocus, handleTextFieldBlur } from "../utils/focus-utils";
+import { spawnChatMessage } from "./chat-message";
+
+const isMobile = AFRAME.utils.device.isMobile();
+
+class InWorldChatBox extends Component {
+  static propTypes = {
+    discordBridges: PropTypes.array,
+    onSendMessage: PropTypes.func,
+    onObjectCreated: PropTypes.func
+  };
+
+  state = {
+    pendingMessage: ""
+  };
+
+  sendMessage = e => {
+    e.preventDefault();
+    if (this.props.onSendMessage) {
+      this.props.onSendMessage(this.state.pendingMessage);
+    }
+    this.setState({ pendingMessage: "" });
+  };
+
+  render() {
+    const textRows = this.state.pendingMessage.split("\n").length;
+    const pendingMessageTextareaHeight = textRows * 28 + "px";
+    const pendingMessageFieldHeight = textRows * 28 + 20 + "px";
+    const discordSnippet = this.props.discordBridges.map(ch => "#" + ch).join(", ");
+
+    return (
+      <form onSubmit={this.sendMessage}>
+        <div
+          className={classNames({ [styles.messageEntryInRoom]: true, [styles.messageEntryOnMobile]: isMobile })}
+          style={{ height: pendingMessageFieldHeight }}
+        >
+          {this.state.pendingMessage.startsWith("/") && (
+            <ChatCommandHelp matchingPrefix={this.state.pendingMessage.substring(1)} />
+          )}
+          <input
+            id="message-entry-media-input"
+            type="file"
+            style={{ display: "none" }}
+            accept={isMobile ? "image/*" : "*"}
+            multiple
+            onChange={e => {
+              if (this.props.onObjectCreated) {
+                for (const file of e.target.files) {
+                  this.props.onObjectCreated(file);
+                }
+              }
+            }}
+          />
+          <label
+            htmlFor="message-entry-media-input"
+            title={"Upload"}
+            className={classNames([
+              styles.messageEntryButton,
+              styles.messageEntryButtonInRoom,
+              styles.messageEntryUpload
+            ])}
+          >
+            <i>
+              <FontAwesomeIcon icon={isMobile ? faCamera : faPlus} />
+            </i>
+          </label>
+          <textarea
+            style={{ height: pendingMessageTextareaHeight }}
+            className={classNames([styles.messageEntryInput, styles.messageEntryInputInRoom, "chat-focus-target"])}
+            value={this.state.pendingMessage}
+            rows={textRows}
+            onFocus={e => {
+              handleTextFieldFocus(e.target);
+            }}
+            onBlur={() => {
+              handleTextFieldBlur();
+            }}
+            onChange={e => {
+              e.stopPropagation();
+              this.setState({ pendingMessage: e.target.value });
+            }}
+            onKeyDown={e => {
+              if (e.key === "Enter" && !e.ctrlKey && !e.shiftKey) {
+                this.sendMessage(e);
+              } else if (e.key === "Enter" && e.ctrlKey) {
+                spawnChatMessage(e.target.value);
+                this.setState({ pendingMessage: "" });
+                e.target.blur();
+              } else if (e.key === "Escape") {
+                e.target.blur();
+              }
+            }}
+            placeholder={this.props.discordBridges.length ? `Send to room and ${discordSnippet}...` : "Send to room..."}
+          />
+          <button
+            className={classNames([styles.messageEntrySpawn])}
+            onClick={() => {
+              if (this.state.pendingMessage.length > 0) {
+                spawnChatMessage(this.state.pendingMessage);
+                this.setState({ pendingMessage: "" });
+              } else {
+                this.pushHistoryState("modal", "create");
+              }
+            }}
+          />
+          <button
+            type="submit"
+            className={classNames([
+              styles.messageEntryButton,
+              styles.messageEntryButtonInRoom,
+              styles.messageEntrySubmit
+            ])}
+          >
+            <i>
+              <FontAwesomeIcon icon={faPaperPlane} />
+            </i>
+          </button>
+        </div>
+      </form>
+    );
+  }
+}
+
+export default InWorldChatBox;
